Rename launchScreenVC to AnimationVC, add launch screen animation.
