+++ BetterReads/BetterReads/Storyboards/Base.lproj/LaunchScreen.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16096" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" launchScreen="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="01J-lp-oVM">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16097" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" launchScreen="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="01J-lp-oVM">
     <device id="retina6_1" orientation="portrait" appearance="light"/>
     <dependencies>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
@@ -15,11 +15,11 @@
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="BetterReads-Graphics_LaunchScreen" translatesAutoresizingMaskIntoConstraints="NO" id="rGE-KT-MkS" userLabel="Blob Background Image">
-                                <rect key="frame" x="-91.5" y="248" width="597" height="400"/>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="BetterReads-Logo_Icon" translatesAutoresizingMaskIntoConstraints="NO" id="rGE-KT-MkS" userLabel="Blob Background Image">
+                                <rect key="frame" x="32" y="273" width="350" height="350"/>
                                 <constraints>
-                                    <constraint firstAttribute="height" constant="400" id="Sph-wr-3j6"/>
-                                    <constraint firstAttribute="width" constant="597" id="ntb-oz-EjI"/>
+                                    <constraint firstAttribute="width" constant="350" id="JgG-vT-l0Z"/>
+                                    <constraint firstAttribute="width" secondItem="rGE-KT-MkS" secondAttribute="height" multiplier="1:1" id="sj4-dr-6Rc"/>
                                 </constraints>
                             </imageView>
                         </subviews>
@@ -37,6 +37,6 @@
         </scene>
     </scenes>
     <resources>
-        <image name="BetterReads-Graphics_LaunchScreen" width="1000" height="671"/>
+        <image name="BetterReads-Logo_Icon" width="2134" height="2134"/>
     </resources>
 </document>
